add HMM Example to test

+++ src/absyn/MapExpr.cpp
@@ -38,9 +38,9 @@ void MapExpr::addMap(Expr* from, Expr* to) {
 void MapExpr::print(FILE* file, int indent) {
   fprintf(file, "%*s(MapExpr:\n", indent, "");
   for (size_t i = 0; i < mapSize(); i++) {
-    fprintf(file, "%*s:from#%zu\n", indent + 2, "", i);
+    fprintf(file, "%*s:from#%u\n", indent + 2, "", (unsigned)i);
     getFrom(i)->print(file, indent + 4);
-    fprintf(file, "%*s:to#%zu\n", indent + 2, "", i);
+    fprintf(file, "%*s:to#%u\n", indent + 2, "", (unsigned)i);
     getTo(i)->print(file, indent + 4);
   }
   fprintf(file, "%*s)\n", indent, "");
